update to solve 250 MB error
